Start of converting to informix fml

+++ src/main/java/no/ntnu/bachelor_group3/dataaccessevaluation/Data/Checkpoint.java
@@ -25,7 +25,7 @@ public class Checkpoint {
 
     private CheckpointType type;
 
-    private double cost;
+    private float cost;
 
     private String location;
 
@@ -51,11 +51,11 @@ public class Checkpoint {
         this.checkpoint_id = counter++;
         this.type = type;
         switch (type) {
-            case Collected -> this.cost = 1;
-            case ReceivedFinalTerminal, ReceivedFirstTerminal -> this.cost = 1.25;
-            case LoadedOnCar -> this.cost = 1.5;
-            case UnderDelivery -> this.cost = 1.8;
-            case Delivered -> this.cost = 2;
+            case Collected -> this.cost = 1f;
+            case ReceivedFinalTerminal, ReceivedFirstTerminal -> this.cost = 1.25f;
+            case LoadedOnCar -> this.cost = 1.5f;
+            case UnderDelivery -> this.cost = 1.8f;
+            case Delivered -> this.cost = 2f;
         }
     }
 
@@ -65,11 +65,11 @@ public class Checkpoint {
         this.type = type;
         this.terminal = terminal;
         switch (type) {
-            case Collected -> this.cost = 1;
-            case ReceivedFinalTerminal, ReceivedFirstTerminal -> this.cost = 1.25;
-            case LoadedOnCar -> this.cost = 1.5;
-            case UnderDelivery -> this.cost = 1.8;
-            case Delivered -> this.cost = 2;
+            case Collected -> this.cost = 1f;
+            case ReceivedFinalTerminal, ReceivedFirstTerminal -> this.cost = 1.25f;
+            case LoadedOnCar -> this.cost = 1.5f;
+            case UnderDelivery -> this.cost = 1.8f;
+            case Delivered -> this.cost = 2f;
         }
     }
 
@@ -90,7 +90,7 @@ public class Checkpoint {
         return sdf3.format(time);
     }
 
-    public double getCost() {
+    public float getCost() {
         return cost;
     }
